Bugfix for issues in description
Fixed the issue where events log on the etlas was not showing the correct events for the mag contacts. This fix is done in `events.py`, where I removed the extra `update_server_events()` calls which are unnecessary.

Also fixed the bug whereby the timed input events for events management was not working properly, which is done in `eventsMod.py`, where I changed the type of event trigger cb being created for the mag contacts